channel_wait_eof: handle receive window exhaustion
Until now, in blocking mode, if the remote receiving window is
exhausted this function hangs forever as data is not read and the
remote side just keeps waiting for the window to grow before sending
more data.

This patch, makes this function check for that condition and abort
with an error when it happens.